requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
fastmcp>=2.0.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
python-multipart>=0.0.9
websockets>=12.0
openai>=1.0.0
